fix: remove item from cart functionality restored
including updating cart price
minor bug: if last remaining item is removed, check out button does not go away

+++ client/src/components/Profile.js
@@ -10,11 +10,13 @@ const Profile = () => {
   const navigate = useNavigate();
   const dispatch = useDispatch();
   const { loggedIn, currentUser } = useSelector(store => store.sessions);
-  console.log(currentUser)
-  // const [loading, setLoading] = useState(true)
+  const items = useSelector(store => store.items.items)
+  // console.log(currentUser)
+  console.log(items)
 
   useEffect(() => {
     loadSession()
+
   },[])
 
   function handleFormNav () {
@@ -42,20 +44,22 @@ const Profile = () => {
             <div class="max-w-2/4 pt-6 p-2 font-sans text-xl font-semibold text-center">
               Listed Items
                   <div>
-                    {currentUser.items.map(item => {
-                      return (
-                        <div class="p-3 mb-3 w-9/12 m-auto bg-slate-200 rounded border-2 border-black">
-                          <div class="text-lg text-left">
-                            {item.title}
+                    {items.map(item => {
+                      if (item.user_id === currentUser.id) {
+                        return (
+                          <div class="p-3 mb-3 w-9/12 m-auto bg-slate-200 rounded border-2 border-black">
+                            <div class="text-lg text-left">
+                              {item.title}
+                            </div>
+                            <div class="text-lg text-right">
+                            <button id={item.id} onClick={(e) => handleDeleteListing(e)} class="inset-0 bg-white hover:bg-gray-100 text-gray-800 font-semibold py-2 px-4 border border-gray-400 rounded shadow">Delete Listing</button>
+                            <NavLink to={`/items/${item.id}`} href={`/items/${item.id}`} class="font-sans text-lg text-black">
+                              <button onClick={() => handleFormNav()} class="inset-0 bg-white hover:bg-gray-100 text-gray-800 font-semibold py-2 px-4 border border-gray-400 rounded shadow">Go To Listing</button>
+                            </NavLink>
+                            </div>
                           </div>
-                          <div class="text-lg text-right">
-                          <button id={item.id} onClick={(e) => handleDeleteListing(e)} class="inset-0 bg-white hover:bg-gray-100 text-gray-800 font-semibold py-2 px-4 border border-gray-400 rounded shadow">Delete Listing</button>
-                          <NavLink to={`/items/${item.id}`} href={`/items/${item.id}`} class="font-sans text-lg text-black">
-                            <button onClick={() => handleFormNav()} class="inset-0 bg-white hover:bg-gray-100 text-gray-800 font-semibold py-2 px-4 border border-gray-400 rounded shadow">Go To Listing</button>
-                          </NavLink>
-                          </div>
-                        </div>
-                      )
+                        )
+                      }
                     })}
                   </div>
                 <button onClick={() => handleFormNav()} class="inset-0 bg-white hover:bg-gray-100 text-gray-800 font-semibold py-2 px-4 border border-gray-400 rounded shadow">Create Listing</button>
